users js ir deprecated and separetly moved into a new folder for userRoutes
