+++ app.js
@@ -10,6 +10,10 @@ const SequelizeStore = require("connect-session-sequelize")(session.Store);
 const indexRouter = require("./routes/index");
 const usersRouter = require("./routes/users");
 const questionRouter = require("./routes/question");
+const userLogin = require("./routes/userRoutes/userlogin");
+const signupRouter = require("./routes/userRoutes/usersignup");
+const userProfile = require("./routes/userRoutes/userprofile");
+const userLogout = require("./routes/userRoutes/userlogout");
 const { restoreUser } = require("./auth");
 
 const app = express();
@@ -40,18 +44,17 @@ store.sync();
 
 app.use(restoreUser);
 
-app.use("/", indexRouter);
-app.use("/users", usersRouter);
+app.use(indexRouter);
+app.use(userLogin);
+app.use(signupRouter);
+app.use(userProfile);
+app.use(userLogout);
+// app.use("/users", usersRouter);
 app.use("/questions", questionRouter);
 
 // catch 404 and forward to error handler
 app.use(function (req, res, next) {
   next(createError(404));
-});
-//catch user profiles that don't exist
-app.use(function (req, res, next) {
-  console.log(req.params, `HERGHEHSEHHTSEHSEHSEHES`);
-  next();
 });
 
 // error handler
@@ -62,6 +65,10 @@ app.use(function (err, req, res, next) {
 
   // render the error page
   res.status(err.status || 500);
+  console.log();
+  if (req.session.auth) {
+    return res.render("page-not-found-logged");
+  }
   res.render("page-not-found");
 });
