+++ views/page-not-found-logged.pug
@@ -6,11 +6,9 @@ html(lang="en")
     meta(name="viewport", content="width=device-width, initial-scale=1.0")
     title Not Found | StackCode 
   body 
-    p Seem like you got lost!~
+    p Seems like you got lost!~
 include utils.pug
 
 block content
   +validationErrorSummary(errors)
-  p: a(href='/users/login') Login
-  p: a(href='/users/signup') Don't have an account?
-  p: a(href='/') Go Back Home?
+  p: a(href='/') Head Back Home
