+++ views/user-login.pug
@@ -7,7 +7,7 @@ html
 include utils.pug
 block content
   +validationErrorSummary(errors)
-  form(action='/users/login' method='post')
+  form(action="/users/login" method='post')
     input(type='hidden' name='_csrf' value=csrfToken)
     +field('Email Address', 'email', email)
     +field('Password', 'password', null, 'password')
